airflow/requirements.txt
# Airflow Providers (pinned versions)
apache-airflow-providers-postgres>=5.13.0
apache-airflow-providers-redis>=3.5.0
apache-airflow-providers-docker>=3.10.0

# Database drivers
psycopg2-binary>=2.9.9

# dbt integration
dbt-core>=1.7.0,<2.0.0
dbt-postgres>=1.7.0,<2.0.0
# For Cosmos integration (optional but recommended)
astronomer-cosmos[dbt-postgres]>=1.3.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Development tools
pytest>=8.0.0
black>=24.0.0
flake8>=7.0.0

# Fake data
faker

google-cloud-bigquery>=3.0.0
google-cloud-bigquery-storage>=2.0.0
pyarrow>=10.0.0
db-dtypes>=1.0.0